+++ core/components/babel/docs/changelog.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Create translated resource children by option
-- Mass unlink/delete translated resources of one resource and their children by option
+- Refresh translated resources with the content of a source resource
+- Mass unlink/delete/refresh translated resources of one resource and their children by option
 
 ## [3.4.0] - 2024-10-13